requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/test.txt requirements/test.in
#
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   pynacl
click==8.1.8
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
code-annotations==2.2.0
    # via -r requirements/test.in
coverage[toml]==7.7.1
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-model-utils
    #   django-waffle
    #   edx-django-utils
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-model-utils==5.0.0
    # via -r requirements/base.txt
django-waffle==4.2.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
dnspython==2.7.0
    # via
    #   -r requirements/base.txt
    #   pymongo
edx-django-utils==7.2.0
    # via -r requirements/base.txt
edx-opaque-keys==2.12.0
    # via -r requirements/base.txt
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via code-annotations
markupsafe==3.0.2
    # via jinja2
newrelic==10.7.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
openedx-atlas==0.6.2
    # via -r requirements/base.txt
packaging==24.2
    # via pytest
pbr==6.1.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.5.0
    # via pytest
psutil==7.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pymongo==4.11.3
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==8.3.5
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.10.0
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.2
    # via code-annotations
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.13.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys

# The following packages are considered to be unsafe in a requirements file:
# setuptools
